+++ src/pages/Subjects_Pages.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function Subjects() {
+export default function Subjects_Pages() {
   return <div>Subjects Page</div>;
 }